+++ packages/domain-iam/src/services/authorization/iam.authorization.service.ts
@@ -112,7 +112,6 @@ export class IAMAuthorizationService<
     let newRefreshToken: string | undefined;
     let tokenContent: DecodedTokenContent<UserTokenEnityFields> | undefined;
     try {
-      logger.info('=> HERE');
       const tokenRes = await tokenManager.verifyAccessToken(authToken, {
         deleteFromStoreIfExpired: true,
         identifierDataField,